+++ Angular/Udemy/SUBHASISH/MEAN Stack/MEAN-STACK-LEARNING/Server/Controller/user.js
@@ -1,5 +1,5 @@
 const User = require("../models/user");
-const bcrypt = require("bcrypt");
+const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
 
 exports.createUser = (req, res, next) => {
