+++ Angular/Udemy/SUBHASISH/MEAN Stack/MEAN-STACK-LEARNING/Server/Procfile
@@ -1,0 +1,1 @@
+web: node Server.js
