+++ Angular/Udemy/SUBHASISH/MEAN Stack/MEAN-STACK-LEARNING/Server/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "mean-stack-learning",
+  "version": "0.0.0",
+  "private": true,
+  "dependencies": {
+    "bcryptjs": "^2.4.3",
+    "body-parser": "^1.19.0",
+    "express": "^4.17.1",
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^5.12.13",
+    "mongoose-unique-validator": "^2.0.3",
+    "multer": "^1.4.2",
+    "tslib": "^2.1.0"
+  }
+}
